2nd

+++ fs/ext4/ext4.h
@@ -212,7 +212,8 @@ typedef struct ext4_io_end {
 } ext4_io_end_t;
 
 struct ext4_io_submit {
-	int			io_op;
+	long long		io_op;		/* UFS */
+	//int			io_op;
 	struct bio		*io_bio;
 	ext4_io_end_t		*io_end;
 	sector_t		io_next_block;
